requirements.txt
airtable-python-wrapper==0.13.0
appnope==0.1.2
astroid==2.7.3
backcall==0.2.0
boto3==1.18.36
botocore==1.21.36
branca==0.4.2
CacheControl==0.12.6
cachetools==4.1.1
certifi==2021.5.30
cffi==1.14.1
chardet==3.0.4
charset-normalizer==2.0.4
click==8.0.1
click-plugins==1.1.1
cligj==0.7.2
cryptography==3.3.2
cycler==0.10.0
decorator==5.0.9
Flask==1.1.2
geographiclib==1.52
geopy==2.2.0
google-api-core==1.22.1
google-auth==1.20.1
google-cloud-core==1.4.1
google-cloud-datastore==1.15.1
google-cloud-logging==1.15.1
google-cloud-ndb==1.6.0
googleapis-common-protos==1.52.0
googlemaps==4.5.3
grpcio==1.31.0
gunicorn==20.0.4
idna==3.2
importlib-metadata==4.8.1
ipython==7.27.0
iso3166==1.0.1
isodate==0.6.0
isort==5.9.3
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
mapbox==0.16.1
mapboxcli==0.8.0
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mccabe==0.6.1
mgrs==1.4.2
msgpack==1.0.2
numpy==1.21.2
openrouteservice==2.3.3
packaging==21.0
pandas==1.3.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==2.3.0
polyline==1.4.0
prompt-toolkit==3.0.20
protobuf==3.13.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.10.0
pylint==2.10.2
pymemcache==3.4.0
pyngrok==5.0.5
pyparsing==2.4.7
pyproj==3.2.1
python-dateutil==2.8.2
python-dotenv==0.19.0
python-telegram-bot==12.8
pytz==2021.3
PyYAML==5.4.1
rdflib==6.0.0
redis==3.5.3
regex==2021.4.4
requests==2.26.0
rsa==4.7
s3transfer==0.5.0
scikit-learn==0.24.2
scipy==1.7.1
Shapely==1.7.1
six==1.16.0
sklearn==0.0
sortedcontainers==2.4.0
SPARQLWrapper==1.8.5
threadpoolctl==2.2.0
toml==0.10.2
tornado==6.0.4
tqdm==4.62.2
traitlets==5.1.0
typed-ast==1.4.3
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.5.0
